refactor(webui): isolate transcription socket handling
